% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{alias}
\alias{alias}
\title{Create an alias for an object whose help page can be viewed with .?obj.}
\usage{
alias(obj, full.name = NULL, comment = "alias")
}
\arguments{
\item{obj}{An R object.}

\item{full.name}{The full qualified name of the original object whose help
page should be retained as \code{pkg::name}. If \code{NULL} (by default), use \code{obj}.}

\item{comment}{A comment to place in \code{obj}.}
}
\value{
The original \code{obj} with the \code{comment} attribute set or replaced with \verb{comment =} plus a \code{src}attribute set to \verb{full.name =}
}
\description{
Create an alias for an object whose help page can be viewed with .?obj.
}
\examples{
# Say you prefer is.true() similar to is.na() or is.null()
# but R provides isTRUE().
is.true <- alias(isTRUE)
# This way, you still got access to the right help page for is.true()
\dontrun{
library(svMisc)
.?is.true
}
}
