% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_funs.R
\name{all_funs}
\alias{all_funs}
\title{Find all functions in an Expression}
\usage{
all_funs(expr, max.names = -1L, unique = TRUE, exclude.names = NULL)
}
\arguments{
\item{expr}{an expression or call from which the name of the function are to
be extracted.}

\item{max.names}{the maximum number of names to be returned. -1 indicates no limit (other than vector size limits).}

\item{unique}{a logical value which indicates whether duplicate names should be removed from the value.}

\item{exclude.names}{a character vector with names to exclude, or \code{NULL} for none.}
}
\value{
A character vector with the extracted function names.
}
\description{
Return a character vector containing the name of all the functions in an
expression or a call.
}
\details{
The c code is adapted from base R code \code{do_allnames()} (the later one allows
to extract either variables, or variables + functions, but not functions
alone).
}
\examples{
ff <- ~z(x, y, z, TRUE, "test", l = 4) + (y(z, x(l)) + y(2))
all.vars(ff)
all.names(ff, unique = TRUE)
all_funs(ff)
all_funs(ff, unique = FALSE)
all_funs(ff, exclude.names = "~")
}
