% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_text_lang.R
\name{get_text_lang}
\alias{get_text_lang}
\title{Translate text messages in a different language than the one currently defined in the R session}
\usage{
get_text_lang(
  ...,
  lang = getOption("data.io_lang", default = Sys.getenv("LANGUAGE", unset = "en")),
  domain = NULL,
  trim = TRUE
)
}
\arguments{
\item{...}{one of more character vectors}

\item{lang}{the desired language (usually in two letters, e.g., "en", "fr", "de", etc.)}

\item{domain}{the 'domain' for the translation, a character string or \code{NULL}; see \code{\link[base:gettext]{base::gettext()}} for more details.}

\item{trim}{logical indicating if the white space trimming should happen. \code{trim = FALSE} may be needed for compiled code (C / C++) messages which often end with \verb{\\n}.}
}
\value{
a character vector with translated messages
}
\description{
Translation messages are obtained with \code{\link[base:gettext]{base::gettext()}}. But, there is no way to specify that one needs translated messages in a different language than the current one in R. This function adds a \verb{lang=} argument that allows to do so.
}
\examples{
old_lang <- Sys.setLanguage("fr") # Switch to French for R language

# Retreive strings in same language
get_text_lang("empty model supplied", "incompatible dimensions",
 domain="R-stats", lang = "fr")

# Retrieve strings in different languages
get_text_lang("empty model supplied", "incompatible dimensions",
  domain="R-stats", lang = "en")
get_text_lang("empty model supplied", "incompatible dimensions",
  domain="R-stats", lang = "de")

# Try to get strings translated in an unknown language (just return the strings)
get_text_lang("empty model supplied", "incompatible dimensions",
  domain="R-stats", lang = "xx")

Sys.setLanguage(old_lang) # Restore original language
rm(old_lang)
}
