\name{isJGR}
\alias{isJGR}

\title{ Is R running with the JGR GUI? }
\description{
  Determine if the R GUI is JGR.
}

\usage{
isJGR()
}

\value{
  \code{TRUE} if the R GUI is JGR, \code{FALSE} otherwise.
}

\author{ Philippe Grosjean <phgrosjean@sciviews.org> }

\seealso{ \code{\link{isMac}}, \code{\link{isAqua}}, \code{\link{isRgui}} }

\examples{
isJGR()
}

\keyword{ utilities }

\concept{ OS system platform }
