\name{isMac}
\alias{isMac}

\title{ Is it a Macintosh platform? }
\description{
  This is a shortcut function to determine if R is currently running on a
  Macintosh platform, i.e., it installs mac binary packages.
}
\usage{
isMac()
}
\value{
  \code{TRUE} if the platform is Macintosh, \code{FALSE} otherwise.
}
\author{ Philippe Grosjean <phgrosjean@sciviews.org> }

\seealso{ \code{\link{isWin}}, \code{\link{isAqua}} }

\examples{
isMac()
}
\keyword{ utilities }
