\name{isWin}
\alias{isWin}

\title{ Is it a Windows platform? }
\description{
  Determine if the platform is Windows.
}
\usage{
isWin()
}
\value{
  TRUE if the platform is Windows, FALSE otherwise.
}
\author{ Philippe Grosjean <phgrosjean@sciviews.org> }

\seealso{ \code{\link{isSDI}} }
\examples{
	isWin()
}
\keyword{ utilities }
