% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{package}
\alias{package}
\title{A very silent and multipackage require() function}
\usage{
package(..., warn = TRUE)
}
\arguments{
\item{...}{The name of one or several R packages to load (character strings).}

\item{warn}{If \code{TRUE}, issue a warning if one or several packages are not
loaded.}
}
\value{
\code{TRUE} if all packages are loaded correctly, \code{FALSE} otherwise.
This function is designed to concisely and quitely indicate package
requirements in GUI menu or other GUI actions.
}
\description{
This function loads one or several R packages as silently as
possible and it returns \code{TRUE} only if all packages are loaded successfully.
If at least one loading fails, a short message is printed.
}
\examples{
# This should work...
if (all(package("tools", "methods"))) message("Fine!")
# ... but this not
if (!all(package("tools", "badname", warn = FALSE))) message("Not fine!")
}
\seealso{
\code{\link[=require]{require()}}, \code{\link[=library]{library()}}
}
\concept{
package requirement and loading
}
\keyword{utilities}
