% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data_dot.R
\name{prepare_data_dot}
\alias{prepare_data_dot}
\alias{prepare_data_dot2}
\alias{recall_with_data_dot}
\alias{recall_with_data_dot2}
\title{Prepare a data_dot function}
\usage{
prepare_data_dot(x, is_top_call = TRUE)

prepare_data_dot2(x, y, is_top_call = TRUE)

recall_with_data_dot(
  call,
  arg = ".data",
  value = quote((.)),
  env = parent.frame(2L),
  abort_msg = gettextf("`\%s` must be a 'data.frame'.", arg),
  abort_msg2 = gettext("Implicit data-dot (.) not permitted"),
  abort_msg3 = gettext("Data-dot mechanism activated, but no `.` object found."),
  abort_msg4 =
    gettextf("Define `.` before calling this function, or provide `\%s=` explicitly.",
    arg),
  abort_msg5 = gettextf("See {.help svMisc::data_dot_mechanism} for more infos."),
  abort_frame = parent.frame()
)

recall_with_data_dot2(
  call,
  arg = "x",
  arg2 = "y",
  value = quote((.)),
  env = parent.frame(2L),
  abort_msg = gettextf("`\%s` and `\%s` must both be 'data.frame'.", arg, arg2),
  abort_msg2 = gettext("Implicit data-dot (.) not permitted"),
  abort_msg3 = gettext("Data-dot mechanism activated, but no `.` object found."),
  abort_msg4 =
    gettextf("Define `.` before calling this function, or provide `\%s=` explicitly.",
    arg),
  abort_msg5 = gettextf("See {.help svMisc::data_dot_mechanism} for more infos."),
  abort_frame = parent.frame()
)
}
\arguments{
\item{x}{An argument to check.}

\item{is_top_call}{A logical indicating if this is a top-level call (\code{TRUE} by
default) that should be focused in the call stack in case of an error.}

\item{y}{A second argument.}

\item{call}{A call object, usually a function call. Could be omitted, and in
this case, \code{sys.call()} is invoked.}

\item{arg}{The name of the argument to inject, usually '.data' (default). For
\code{prepare_data_dot2()}, it is \code{x} by default}

\item{value}{The value to inject, usually the symbol \code{.} (default).}

\item{env}{The environment where the evaluation of the data-dot-injected call
should be evaluated (by default, \code{parent.frame(2L)}, should rarely be
changed).}

\item{abort_msg}{The message to use in case the '.data' argument is wrong.}

\item{abort_msg2}{An additional message to append to the error message in
case data-dot-injection is not permitted (when
\code{.SciViews.implicit.data.dot != TRUE}, see example).}

\item{abort_msg3}{An error message when \code{.} is not found.}

\item{abort_msg4}{An additional explanation when \code{.}is not found.}

\item{abort_msg5}{A hint to read the documentation of the data-dot mechanism.}

\item{abort_frame}{The environment to use for the error message, by default,
the caller environment (should rarely be changed).}

\item{arg2}{The name of the second argument, \code{y} by default.}
}
\value{
\code{TRUE} if the preparation is correct for \code{prepare_dzata_dot()},
\code{FALSE} otherwise. The result from evaluating the data-dot-injected call for
\code{recall_with_data_dot()}.
}
\description{
Prepare a function that uses the data-dot mechanism. In case the
argument (usually, \code{.data = (.)}) is missing or is not a data frame in a call
to a "data-dot" function, it is recalled after injection \code{.} as first
argument.
}
\details{
The call is \strong{not} checked if it is a correct function call. When called
from within a function, passing \code{\link[=sys.call]{sys.call()}} as \code{call}, it should be always
correct.
\code{prepare_data_dot2()} and \code{recall_with_data_dot2()} are similar, but for
functions that have two first arguments that must be data frames, generally
called \code{x}and \code{y}.
}
\examples{
# Use this (rename) function to get extra-info in the error message about the
# data-dot mechanism automatically
stop <- stop_

# Here is how you create a data-dot function
my_subset <- function(.data = (.), i, j) {
  if (!prepare_data_dot(.data))
    return(recall_with_data_dot())

  if (!is.numeric(i))
    stop_("{.arg i} must be numeric") # Use this function
  if (i == 1)
    message(".env has ", paste(names(.env), collapse = ", "))
  .data[i, j]
}

dtf1 <- data.frame(x = 1:3, y = 4:6)
# The message shows the objects available in the function environment
my_subset(dtf1, 1, 'y')
# This is wrong
try(my_subset(dtf1, 'y'))
.= dtf1
my_subset(2, 'y')
# Error message with indication that the data-dot mechanism is activated
try(my_subset('y'))
# Data-dot mechanism not activate, but dot object used
try(my_subset(., 'y'))
# Wrong .data=
try(my_subset(.data = 'y'))

# When data-dot is not permitted...
.SciViews.implicit.data.dot <- FALSE
try(my_subset(2, 'y'))
rm(.SciViews.implicit.data.dot)

# When `.` is not found...
rm(.)
try(my_subset(2, 'y'))
}
\seealso{
\link{data_dot_mechanism}
}
