% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retarget.R
\name{retarget}
\alias{retarget}
\title{Retarget a formula to a different environment}
\usage{
retarget(formula, env = parent.frame())
}
\arguments{
\item{formula}{A formula (or a quosure)}

\item{env}{The environment to which the formula should be retargeted.
Defaults to the parent frame.}
}
\value{
The formula with the new environment set.
}
\description{
Change the environment of a formula to a specified environment.
}
\examples{
# A rather inefficient way to build a formula... for the sake of the demo!
make_formula <- function(x)
  as.formula(x)
f <- make_formula("x ~ log(y) + z")
f
f <- retarget(f)
f
# OK, but the environment associated to this formula is...
# the environment of the function:
rlang::f_env(f)
}
