% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_clipboard.R
\name{source_clipboard}
\alias{source_clipboard}
\alias{sourceClipboard}
\title{Source code from the clipboard}
\usage{
source_clipboard(primary = TRUE, ...)

sourceClipboard(primary = TRUE, ...)
}
\arguments{
\item{primary}{Only valid on *nix: read the primary (or secondary) clipboard.}

\item{...}{Further parameters passed to \code{\link[=source]{source()}}.}
}
\value{
Same result as \code{\link[=source]{source()}}.
}
\description{
This function reads R code from the clipboard, and then source
it. Clipboard is managed correctly depending on the OS (Windows, MacOS, or
*nix)
}
\seealso{
\code{\link[=source]{source()}}, \code{\link[=file]{file()}}
}
\concept{Source code from clipboard}
\keyword{IO}
