\encoding{UTF-8}

\name{SVMisc-package}
\alias{SVMisc-package}
\alias{SVMisc}

\docType{package}

\title{
\packageTitle{SVMisc}
}

\description{
\packageDescription{SVMisc}
}

\details{

\tabular{ll}{
Package: \tab SVMisc\cr
Type: \tab Package\cr
%Version: \tab 1.0\cr
%Date: \tab 2015-05-03\cr
License: \tab GPL (>= 2)\cr
}
\packageDESCRIPTION{SVMisc}
}

\author{
\packageAuthor{SVMisc}

Maintainer: \packageMaintainer{SVMisc}
}

\seealso{
\code{\link{sv_cor_prob_matrix}},
\code{\link{sv_cs_nagel}},
\code{\link{sv_exact_polynom_test}},
\code{\link{sv_export_to_SPSS}},
\code{\link{sv_get_levels}},
\code{\link{sv_make_factor}},
\code{\link{sv_make_numeric}},
\code{\link{sv_relevel}}
}
