\name{Sys.userdir}
\alias{Sys.userdir}

\title{ Get the user directory }
\description{
  This function is deprecated in favor of \code{systemDir("user")}!
  \code{Sys.userdir()} retrieves the current user directory.
}

\usage{
Sys.userdir()
}

\value{
  A string with the path to the user directory.
}

\author{ Philippe Grosjean <phgrosjean@sciviews.org> }

\seealso{ \code{\link{systemDir}} }

\keyword{ utilities }
