\name{Sys.userdir}
\alias{Sys.userdir}

\title{ Get the user directory }
\description{
  \code{Sys.userdir()} retrieves the current user directory.
}
\usage{
Sys.userdir()
}

\value{
  A string with the path to the user directory.
}
\author{ Philippe Grosjean <phgrosjean@sciviews.org> }

\seealso{ \code{\link{Sys.tempdir}} }

\examples{
Sys.userdir()
}

\keyword{ utilities }

\concept{ system directories }
