% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_var.R
\name{temp_var}
\alias{temp_var}
\alias{tempvar}
\title{Get an arbitrary name for a temporary variable}
\usage{
temp_var(pattern = ".var")

tempvar(pattern = ".var")
}
\arguments{
\item{pattern}{The prefix for the variable (the rest is a random number).}
}
\value{
A string with the name of a variable.
}
\description{
This function ensures that the variable name is cryptic enough
and is not already used.
}
\examples{
temp_var()
}
\seealso{
\code{\link[=tempfile]{tempfile()}}
}
\concept{temporary variables}
\keyword{utilities}
