\name{userdir}
\alias{userdir}

\title{ Get the user directory }
\description{
  \code{userdir()} retrieves the current user directory.
}
\usage{
    userdir()
}

\value{
  A string with the path to the user directory
}
\author{ Philippe Grosjean <phgrosjean@sciviews.org> }

\seealso{ \code{\link{tempdirWin}} }
\examples{
	userdir()
}
\keyword{ utilities }
