% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gettext.R
\name{gettext}
\alias{gettext}
\alias{gettextf}
\alias{test_gettext_lang}
\title{Translate text messages in a different language than the one currently defined in the R session}
\usage{
gettext(
  ...,
  lang = getOption("data.io_lang", default = Sys.getenv("LANGUAGE", unset = "en")),
  domain = NULL,
  trim = TRUE
)

gettextf(
  fmt,
  ...,
  lang = getOption("data.io_lang", default = Sys.getenv("LANGUAGE", unset = "en")),
  domain = NULL,
  trim = TRUE
)

test_gettext_lang(
  lang = getOption("data.io_lang", default = Sys.getenv("LANGUAGE", unset = "en"))
)
}
\arguments{
\item{...}{one of more character vectors}

\item{lang}{the desired language (usually in two letters, e.g., "en", "fr", "de", etc.)}

\item{domain}{the 'domain' for the translation, a character string or \code{NULL}; see \code{\link[base:gettext]{base::gettext()}} for more details.}

\item{trim}{logical indicating if the white space trimming should happen. \code{trim = FALSE} may be needed for compiled code (C / C++) messages which often end with \verb{\\n}.}

\item{fmt}{a character vector of format strings, each of up to 8192 bytes.}
}
\value{
a character vector with translated messages
}
\description{
Translation messages are obtained with \code{\link[base:gettext]{base::gettext()}}. But, there is no way to specify that one needs translated messages in a different language than the current one in R. This function adds a \verb{lang=} argument that allows to do so.
}
\examples{
old_lang <- Sys.setLanguage("fr") # Switch to French for R language

# Retrieve strings in same language
gettext("empty model supplied", "incompatible dimensions",
 domain="R-stats", lang = "fr")

# Retrieve strings in different languages
gettext("empty model supplied", "incompatible dimensions",
  domain="R-stats", lang = "en")
gettext("empty model supplied", "incompatible dimensions",
  domain="R-stats", lang = "de")

# Try to get strings translated in an unknown language (just return the strings)
gettext("empty model supplied", "incompatible dimensions",
  domain="R-stats", lang = "xx")

# Test with some translations from the svMisc package itself:
test_gettext_lang()
test_gettext_lang("fr")
test_gettext_lang("en")

Sys.setLanguage(old_lang) # Restore original language
rm(old_lang)
}
