\name{isWin}
\alias{isWin}

\title{ Is it a Windows platform? }
\description{
  Shortcut to determine if the platform is Windows.
}

\usage{
isWin()
}

\value{
  \code{TRUE} if the platform is Windows, \code{FALSE} otherwise.
}

\author{ Philippe Grosjean <phgrosjean@sciviews.org> }

\seealso{ \code{\link{isMac}}, \code{\link{isRgui}}, \code{\link{isJGR}} }

\examples{
isWin()
}

\keyword{ utilities }

\concept{ OS system platform }
