% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svMisc-package.R
\docType{package}
\name{svMisc-package}
\alias{svMisc-package}
\title{SciViews - Miscellaneous functions}
\description{
Miscellaneous functions for SciViews or general use. The svMisc package
collects together a series of functions that are shared with svXXX packages.
}
\section{Important functions}{

\itemize{
\item \code{\link[=temp_env]{temp_env()}} for unsing a temporary environment attached to the search path,
\item \code{\link[=temp_var]{temp_var()}} create the name of temporary variables,
\item \code{\link[=capture_all]{capture_all()}} to capture R output, errors, warnings and messages,
\item \code{\link[=parse_text]{parse_text()}} to parse any R expression, including partial or incorrect
ones (fails gracefully).
}
}

