\name{svMisc-package}
\alias{svMisc-package}
\alias{svMisc}
\docType{package}

\title{ SciViews GUI API - Miscellaneous functions }
\description{
  The SciViews svMisc package collects together a series of functions that are
  shared across all the svXXX packages.
}
\details{
  \tabular{ll}{
    Package: \tab svMisc\cr
    Type: \tab Package\cr
    Version: \tab 0.9-70\cr
    Date: \tab 2014-03-01\cr
    License: \tab GPL 2 or above, at your convenience\cr
  }
  % TODO: add description of main functions here. Also add examples
}

\author{
  Philippe Grosjean, Romain Francois & Kamil Barton.

  Maintainer: Ph. Grosjean <phgrosjean@sciviews.org>
}

\keyword{ package }
\keyword{ utilities }

\concept{ GUI API miscellaneous functions }
