% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svMisc-package.R
\docType{package}
\name{svMisc-package}
\alias{svMisc-package}
\title{Miscellaneous Functions for 'SciViews::R'}
\description{
The \{svMisc\} package is of general use among SciViews::R, a layer on top of
R, and the tidyverse. This package collects together a series of general
functions to manage a centralized environment for temporary variables, a
progress bar and batch analysis mode, etc.
}
\section{Important functions}{

\itemize{
\item \code{\link[=temp_env]{temp_env()}} for using a temporary environment attached to the search path,
\item \code{\link[=temp_var]{temp_var()}} create the name of temporary variables,
\item \code{\link[=capture_all]{capture_all()}} to capture R output, errors, warnings and messages,
\item \code{\link[=parse_text]{parse_text()}} to parse any R expression, including partial or incorrect
ones (fails gracefully).
}
}

